requirements.txt
pytesseract
PyMuPDF
numpy
opencv-python
pandas
transformers
sentence-transformers
faiss-cpu